+++ action-src/main.ts
@@ -129,7 +129,9 @@ async function run() {
     const only = getInput('only');
     const onlyChanged = getInput('onlyChanged');
     const externals = getInput('externals');
+    const untraced = getInput('untraced');
     const doNotStart = getInput('doNotStart');
+    const diagnostics = getInput('diagnostics');
     const storybookPort = getInput('storybookPort');
     const storybookUrl = getInput('storybookUrl');
     const storybookBuildDir = getInput('storybookBuildDir');
@@ -155,11 +157,13 @@ async function run() {
       workingDir: maybe(workingDir),
       buildScriptName: maybe(buildScriptName),
       scriptName: maybe(scriptName),
+      diagnostics: maybe(diagnostics),
       exec: maybe(exec),
       skip: maybe(skip),
       only: maybe(only),
       onlyChanged: maybe(onlyChanged),
       externals: maybe(externals),
+      untraced: maybe(untraced),
       doNotStart: maybe(doNotStart),
       storybookPort: maybe(storybookPort),
       storybookUrl: maybe(storybookUrl),