+++ bin-src/lib/getDependentStoryFiles.js
@@ -1,8 +1,9 @@
 import path from 'path';
 
-import { getWorkingDir } from './utils';
+import { getWorkingDir, matchesFile } from './utils';
 import { getRepositoryRoot } from '../git/git';
 import bailFile from '../ui/messages/warnings/bailFile';
+import noCSFGlobs from '../ui/messages/errors/noCSFGlobs';
 
 // Bail whenever one of these was changed
 const GLOBALS = [
@@ -17,7 +18,7 @@ const GLOBALS = [
 // Ignore these while tracing dependencies
 const EXTERNALS = [/^node_modules\//, /\/node_modules\//, /\/webpack\/runtime\//, /^\(webpack\)/];
 
-const isGlobal = (name) => GLOBALS.some((re) => re.test(name));
+const isPackageFile = (name) => GLOBALS.some((re) => re.test(name));
 const isUserModule = ({ id, name, moduleName }) =>
   id !== undefined && id !== null && !EXTERNALS.some((re) => re.test(name || moduleName));
 
@@ -44,15 +45,16 @@ export function normalizePath(posixPath, rootPath, workingDir = '') {
  * the changed git files. The result is a map of Module ID => file path. In the end we'll only send
  * the Module IDs to Chromatic, the file paths are only for logging purposes.
  */
-export async function getDependentStoryFiles(ctx, stats, changedFiles) {
-  const { configDir = '.storybook', staticDir = [] } = ctx.storybook || {};
-  const { storybookBaseDir } = ctx.options;
+export async function getDependentStoryFiles(ctx, stats, statsPath, changedFiles) {
+  const { configDir = '.storybook', staticDir = [], viewLayer } = ctx.storybook || {};
+  const { storybookBaseDir, untraced = [] } = ctx.options;
 
   // Currently we enforce Storybook to be built by the Chromatic CLI, to ensure absolute paths match
   // up between the webpack stats and the git repo root.
   const rootPath = await getRepositoryRoot(); // e.g. `/path/to/project` (always absolute posix)
   const workingDir = getWorkingDir(rootPath, storybookBaseDir); // e.g. `packages/storybook` or empty string
-  const normalize = (posixPath) => normalizePath(posixPath, rootPath, workingDir);
+  const normalize = (posixPath) => normalizePath(posixPath, rootPath, workingDir); // e.g. `src/file.js` (no ./ prefix)
+  const baseName = (name) => normalize(name).replace(/ \+ \d+ modules$/, '');
 
   const storybookDir = normalize(posix(configDir));
   const staticDirs = staticDir.map((dir) => normalize(posix(dir)));
@@ -93,30 +95,50 @@ export async function getDependentStoryFiles(ctx, stats, changedFiles) {
     }
   });
 
-  ctx.log.info(`Found ${Object.keys(csfGlobsByName).length} CSF globs`);
-  ctx.log.info(`Found ${Object.keys(idsByName).length} user modules`);
+  ctx.turboSnap.globs = Object.keys(csfGlobsByName);
+  ctx.turboSnap.modules = Object.keys(idsByName);
+
+  if (ctx.turboSnap.globs.length === 0) {
+    ctx.log.error(noCSFGlobs({ statsPath, storybookDir, viewLayer }));
+    throw new Error('Did not find any CSF globs in preview-stats.json');
+  }
 
   const isCsfGlob = (name) => !!csfGlobsByName[name];
-  const isConfigFile = (name) =>
+  const isStorybookFile = (name) =>
     name && name.startsWith(`${storybookDir}/`) && !storiesEntryFiles.includes(name);
   const isStaticFile = (name) => staticDirs.some((dir) => name && name.startsWith(`${dir}/`));
 
+  ctx.untracedFiles = [];
+  function untrace(filepath) {
+    if (untraced.some((glob) => matchesFile(glob, filepath))) {
+      ctx.untracedFiles.push(filepath);
+      return false;
+    }
+    return true;
+  }
+
+  const tracedFiles = changedFiles.filter(untrace);
   const changedCsfIds = new Set();
   const checkedIds = {};
   const toCheck = [];
 
-  let bail = changedFiles.find(isGlobal);
+  const changedPackageFile = changedFiles.find(isPackageFile);
+  if (changedPackageFile) ctx.turboSnap.bailReason = { changedPackageFile };
 
   function shouldBail(name) {
-    if (isConfigFile(name) || isStaticFile(name)) {
-      bail = name;
+    if (isStorybookFile(name)) {
+      ctx.turboSnap.bailReason = { changedStorybookFile: baseName(name) };
+      return true;
+    }
+    if (isStaticFile(name)) {
+      ctx.turboSnap.bailReason = { changedStaticFile: baseName(name) };
       return true;
     }
     return false;
   }
 
   function traceName(normalizedName) {
-    if (bail || isCsfGlob(normalizedName)) return;
+    if (ctx.turboSnap.bailReason || isCsfGlob(normalizedName)) return;
     if (shouldBail(normalizedName)) return;
 
     const id = idsByName[normalizedName];
@@ -131,21 +153,20 @@ export async function getDependentStoryFiles(ctx, stats, changedFiles) {
     }
   }
 
-  changedFiles.forEach(traceName);
+  tracedFiles.forEach(traceName);
   while (toCheck.length > 0) {
     const id = toCheck.pop();
     checkedIds[id] = true;
-    reasonsById[id].forEach(traceName);
+    reasonsById[id].filter(untrace).forEach(traceName);
   }
 
-  if (bail) {
-    ctx.log.warn(bailFile(bail));
-    return false;
+  if (ctx.turboSnap.bailReason) {
+    ctx.log.warn(bailFile(ctx));
+    return null;
   }
 
   return stats.modules.reduce((acc, mod) => {
-    if (changedCsfIds.has(mod.id))
-      acc[String(mod.id)] = normalize(mod.name).replace(/ \+ \d+ modules$/, '');
+    if (changedCsfIds.has(mod.id)) acc[String(mod.id)] = baseName(mod.name);
     return acc;
   }, {});
 }