+++ bin-src/lib/getOptions.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import { parse } from 'url';
 
+import dependentOption from '../ui/messages/errors/dependentOption';
 import duplicatePatchBuild from '../ui/messages/errors/duplicatePatchBuild';
 import incompatibleOptions from '../ui/messages/errors/incompatibleOptions';
 import invalidExitOnceUploaded from '../ui/messages/errors/invalidExitOnceUploaded';
@@ -18,11 +19,13 @@ import inferredOptions from '../ui/messages/info/inferredOptions';
 import getStorybookConfiguration from './getStorybookConfiguration';
 
 const takeLast = (input) => (Array.isArray(input) ? input[input.length - 1] : input);
+const ensureArray = (input) => (Array.isArray(input) ? input : [input]);
 
 const resolveHomeDir = (filepath) =>
   filepath && filepath.startsWith('~') ? path.join(process.env.HOME, filepath.slice(1)) : filepath;
 
 const trueIfSet = (value) => (value === '' ? true : value);
+const undefinedIfEmpty = (array) => (array.length ? array : undefined);
 
 export default async function getOptions({ argv, env, flags, log, packageJson }) {
   const fromCI = !!flags.ci || !!process.env.CI;
@@ -34,7 +37,8 @@ export default async function getOptions({ argv, env, flags, log, packageJson })
 
     only: flags.only,
     onlyChanged: trueIfSet(flags.onlyChanged),
-    externals: flags.externals,
+    untraced: undefinedIfEmpty(ensureArray(flags.untraced)),
+    externals: undefinedIfEmpty(ensureArray(flags.externals)),
     list: flags.list,
     fromCI,
     skip: trueIfSet(flags.skip),
@@ -119,6 +123,14 @@ export default async function getOptions({ argv, env, flags, log, packageJson })
 
   if (options.only && options.onlyChanged) {
     throw new Error(invalidSingularOptions(['--only', '--only-changed']));
+  }
+
+  if (options.untraced && !options.onlyChanged) {
+    throw new Error(dependentOption('--untraced', '--only-changed'));
+  }
+
+  if (options.externals && !options.onlyChanged) {
+    throw new Error(dependentOption('--externals', '--only-changed'));
   }
 
   // No need to start or build Storybook if we're going to fetch from a URL