+++ bin-src/lib/parseArgs.js
@@ -22,9 +22,10 @@ export default function parseArgs(argv) {
       --branch-name <branch>  Override the branch name. Only meant to be used for unsupported CI integrations and fixing cross-fork PR comparisons. Also accepts <owner>:<branch> format.
       --exit-once-uploaded [branch]  Exit with 0 once the built version has been published to Chromatic. Only for [branch], if specified. Globs are supported via picomatch.
       --exit-zero-on-changes [branch]  If all snapshots render but there are visual changes, exit with code 0 rather than the usual exit code 1. Only for [branch], if specified. Globs are supported via picomatch.
-      --externals <filepath>  When any of these files have changed since the baseline build, ignore --only-changed. Globs are supported via picomatch. This flag can be specified multiple times.
+      --externals <filepath>  Disable TurboSnap when any of these files have changed since the baseline build. Globs are supported via picomatch. This flag can be specified multiple times. Requires --only-changed.
+      --untraced <filepath>  Disregard these files and their dependencies when tracing dependent stories for TurboSnap. Globs are supported via picomatch. This flag can be specified multiple times. Requires --only-changed.
       --ignore-last-build-on-branch <branch>  Do not use the last build on this branch as a baseline if it is no longer in history (i.e. branch was rebased). Globs are supported via picomatch.
-      --only-changed [branch]  Only run stories affected by files changed since the baseline build. Only for [branch], if specified. Globs are supported via picomatch. All other snapshots will be inherited from the prior commit.
+      --only-changed [branch]  Enables TurboSnap: Only run stories affected by files changed since the baseline build. Only for [branch], if specified. Globs are supported via picomatch. All other snapshots will be inherited from the prior commit.
       --patch-build <headbranch...basebranch>  Create a patch build to fix a missing PR comparison.
       --preserve-missing  Treat missing stories as unchanged rather than deleted when comparing to the baseline.
       --skip [branch]  Skip Chromatic tests, but mark the commit as passing. Avoids blocking PRs due to required merge checks. Only for [branch], if specified. Globs are supported via picomatch.
@@ -38,7 +39,8 @@ export default function parseArgs(argv) {
       --junit-report [filepath]  Write build results to a JUnit XML file. {buildNumber} will be replaced with the actual build number. [chromatic-build-{buildNumber}.xml]
       --list  List available stories. This requires running a full build.
       --no-interactive  Don't ask interactive questions about your setup and don't overwrite output. Always true in non-TTY environments.
-      --only <storypath>  Only run a single story or a subset of stories. Story paths typically look like "Path/To/Story". Globs are supported via picomatch. All other snapshots will be inherited from the prior commit. This option implies --preserve-missing.
+      --only <storypath>  Only run a single story or a subset of stories. Story paths typically look like "Path/To/Story". Globs are supported via picomatch. This option implies --preserve-missing.
+      --diagnostics  Write process context information to chromatic-diagnostics.json.
     `,
     {
       argv,
@@ -61,6 +63,7 @@ export default function parseArgs(argv) {
         ignoreLastBuildOnBranch: { type: 'string' },
         only: { type: 'string' },
         onlyChanged: { type: 'string' },
+        untraced: { type: 'string', isMultiple: true },
         externals: { type: 'string', isMultiple: true },
         branchName: { type: 'string' },
         patchBuild: { type: 'string' },
@@ -72,6 +75,7 @@ export default function parseArgs(argv) {
         // Debug options
         ci: { type: 'boolean' },
         debug: { type: 'boolean' },
+        diagnostics: { type: 'boolean' },
         dryRun: { type: 'boolean' },
         junitReport: { type: 'string' },
         list: { type: 'boolean' },