+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# 6.3.0 - 2021-12-22
+
+- [461](https://github.com/chromaui/chromatic-cli/pull/461) Add `--untraced` flag to avoid retesting stories that depend on certain files
+- [479](https://github.com/chromaui/chromatic-cli/pull/479) Add `--diagnostics` flag to write process context data to a file
+- [458](https://github.com/chromaui/chromatic-cli/pull/458) Track `bailReason`, improve TurboSnap messaging and throw on zero CSF globs
+- [482](https://github.com/chromaui/chromatic-cli/pull/482) Fix commit details when using env var and warn if the commit is missing
+
+# 6.2.3 - 2021-12-17
+
+- Avoid optional chaining which breaks in Node 12 (GitHub Actions)
+
+# 6.2.2 - 2021-12-17
+
+- Fix error handling in GraphQL client to not retry mutation on HTTP error
+
+# 6.2.1 - 2021-12-16
+
+- [477](https://github.com/chromaui/chromatic-cli/pull/477) Retry createBuild based on error messages
+- [468](https://github.com/chromaui/chromatic-cli/pull/468) Increase unpack wait timeout to 3 minutes
+- [466](https://github.com/chromaui/chromatic-cli/pull/466) Add workingDirectory input handler for action
+- [465](https://github.com/chromaui/chromatic-cli/pull/465) Remove the need to set a GitHub token
+
+# 6.2.0 - 2021-12-07
+
+- [459](https://github.com/chromaui/chromatic-cli/pull/459) Add --zip flag to upload files as zip archive
+- [463](https://github.com/chromaui/chromatic-cli/pull/463) Fix tests of makeZipFile
+- [447](https://github.com/chromaui/chromatic-cli/pull/447) Add support for passing the same flag multiple times
+
 # 6.1.0 - 2021-11-29
 
 - [455](https://github.com/chromaui/chromatic-cli/pull/455) Add `--storybook-base-dir` to support TurboSnap with a prebuilt Storybook originating from a subdirectory